Modularizado App.js
