+++ package.json
@@ -28,7 +28,6 @@
     "react-redux": "^7.2.6",
     "redux": "^4.1.2",
     "redux-thunk": "^2.4.0"
-
   },
   "devDependencies": {
     "@babel/core": "^7.12.9"
